+++ src/nes_state/mod.rs
@@ -13,7 +13,7 @@ pub type LocalNesState = RusticNesState;
 #[derive(Clone)]
 pub struct FrameData {
     pub video: Vec<u8>,
-    pub audio: Vec<i16>,
+    pub audio: Vec<f32>,
     pub fps: Fps,
 }
